requirements.txt
streamlit>=1.30.0
ultralytics>=8.0.0
opencv-python>=4.8.0
numpy>=1.23.0
pillow>=9.5.0
